modify my json

+++ models/adam_vpt.py
@@ -140,7 +140,7 @@ class Learner(BaseLearner):
                 inputs, targets = inputs.to(self._device), targets.to(self._device)
                 logits = self._network(inputs)["logits"]
 
-                loss = F.cross_entropy(logits, targets)
+                loss = F.cross_entropy(logits, targets.long())
                 optimizer.zero_grad()
                 loss.backward()
                 optimizer.step()
